requirements.txt
pydantic
pydantic_settings

requests

schedule

fastapi
uvicorn

paho-mqtt

logger
